genai-observability/load-generator/requirements.txt
openai>=1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-openai-v2
